+++ src/js/debugger/src/BodyComponent.tsx
